--- account_cutoff_prepaid/models/__init__.py
@@ -1,3 +1,0 @@
-from . import company
-from . import res_config_settings
-from . import account_cutoff
