--- account_cutoff_prepaid/readme/CONFIGURE.rst
@@ -1,5 +1,0 @@
-On the form view of the company, in the *Configuration* tab, there is a *Cut-offs* section where you can configure:
-
-* the *Default Cut-off Journal*, which is the journal in which the cut-off journal items will be generated by default,
-* the *Default Account for Prepaid Revenue*,
-* the *Default Account for Prepaid Expense*.
