+++ account_cutoff_start_end_dates/readme/CONFIGURE.rst
@@ -1,0 +1,1 @@
+Please refer to the **CONFIGURATION** section of the README of the module *account_cutoff_base*.
