+++ account_cutoff_start_end_dates/views/account_cutoff.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <!--
-  Copyright 2013-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+  Copyright 2013-2021 Akretion France (http://www.akretion.com/)
+  @author: Alexis de Lattre <alexis.delattre@akretion.com>
   License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
 <odoo>
@@ -13,15 +14,15 @@
             <field name="cutoff_date" position="after">
                 <field
                     name="forecast"
-                    invisible="'prepaid' not in context.get('cutoff_type', '-')"
+                    invisible="'prepaid' not in context.get('default_cutoff_type', '-')"
                 />
                 <field
                     name="start_date"
-                    invisible="'prepaid' not in context.get('cutoff_type', '-')"
+                    invisible="'prepaid' not in context.get('default_cutoff_type', '-')"
                 />
                 <field
                     name="end_date"
-                    invisible="'prepaid' not in context.get('cutoff_type', '-')"
+                    invisible="'prepaid' not in context.get('default_cutoff_type', '-')"
                 />
             </field>
         </field>
@@ -130,18 +131,13 @@
         <field name="model">account.cutoff.line</field>
         <field name="inherit_id" ref="account_cutoff_base.account_cutoff_line_form" />
         <field name="arch" type="xml">
-            <field name="parent_id" position="after">
-                <field name="move_line_id" />
-                <field name="move_id" />
-                <field name="move_date" />
-            </field>
             <field name="name" position="after">
                 <field name="start_date" />
                 <field name="end_date" />
             </field>
             <field name="cutoff_amount" position="before">
                 <field name="total_days" />
-                <field name="prepaid_days" />
+                <field name="cutoff_days" />
             </field>
         </field>
     </record>
@@ -150,57 +146,12 @@
         <field name="model">account.cutoff.line</field>
         <field name="inherit_id" ref="account_cutoff_base.account_cutoff_line_tree" />
         <field name="arch" type="xml">
-            <field name="parent_id" position="after">
-                <field name="move_id" />
-            </field>
             <field name="analytic_account_id" position="after">
-                <field name="start_date" />
-                <field name="end_date" />
-                <field name="total_days" />
-                <field name="prepaid_days" />
+                <field name="start_date" optional="show" />
+                <field name="end_date" optional="show" />
+                <field name="total_days" optional="show" />
+                <field name="cutoff_days" optional="show" />
             </field>
         </field>
     </record>
-    <record id="account_cutoff_prepaid_expense_action" model="ir.actions.act_window">
-        <field name="name">Prepaid Expense</field>
-        <field name="res_model">account.cutoff</field>
-        <field name="view_mode">tree,form</field>
-        <field name="domain">[('cutoff_type', '=', 'prepaid_expense')]</field>
-        <field name="context">{'cutoff_type': 'prepaid_expense'}</field>
-        <field name="help" type="html">
-            <p class="oe_view_nocontent_create">
-            Click to start preparing a new prepaid expense.
-            </p>
-            <p>
-            This view can be used by accountants in order to collect information about prepaid expenses based on start date and end date. It then allows to generate the corresponding cutoff journal entry in one click.
-            </p>
-        </field>
-    </record>
-    <menuitem
-        id="account_cutoff_prepaid_expense_menu"
-        parent="account_cutoff_base.cutoff_menu"
-        action="account_cutoff_prepaid_expense_action"
-        sequence="25"
-    />
-    <record id="account_cutoff_prepaid_revenue_action" model="ir.actions.act_window">
-        <field name="name">Prepaid Revenue</field>
-        <field name="res_model">account.cutoff</field>
-        <field name="view_mode">tree,form</field>
-        <field name="domain">[('cutoff_type', '=', 'prepaid_revenue')]</field>
-        <field name="context">{'cutoff_type': 'prepaid_revenue'}</field>
-        <field name="help" type="html">
-            <p class="oe_view_nocontent_create">
-            Click to start preparing a new prepaid revenue.
-            </p>
-            <p>
-            This view can be used by accountants in order to collect information about prepaid revenues based on start date and end date. It then allows to generate the corresponding cutoff journal entry in one click.
-            </p>
-        </field>
-    </record>
-    <menuitem
-        id="account_cutoff_prepaid_revenue_menu"
-        parent="account_cutoff_base.cutoff_menu"
-        action="account_cutoff_prepaid_revenue_action"
-        sequence="20"
-    />
 </odoo>
