+++ setup/account_cutoff_start_end_dates/odoo/addons/account_cutoff_start_end_dates
@@ -1,0 +1,1 @@
+../../../../account_cutoff_start_end_dates
